Enable common warnings for all (but dive_core) (#895)
This is a follow-on to #894 that addresses the rest of the directories.
I left dive_core since it's slightly trickier; this will be addressed in
a future PR.

To harmonize the code base, I encorporated -Wextra into the common
warnings as well.

+++ gfxr_dump_resources/gfxr_dump_resources.cpp
@@ -74,7 +74,7 @@ bool SaveAsJsonFile(const std::vector<DumpEntry>& dumpables, const char* filenam
     out << "  },\n";
 
     out << "  \"BeginCommandBuffer\": [";
-    for (int i = 0; i < dumpables.size(); ++i)
+    for (size_t i = 0; i < dumpables.size(); ++i)
     {
         const DumpEntry& entry = dumpables[i];
         if (i != 0)
@@ -86,7 +86,7 @@ bool SaveAsJsonFile(const std::vector<DumpEntry>& dumpables, const char* filenam
     out << "],\n";
 
     out << "  \"RenderPass\": [";
-    for (int i = 0; i < dumpables.size(); ++i)
+    for (size_t i = 0; i < dumpables.size(); ++i)
     {
         const DumpEntry& entry = dumpables[i];
         if (i != 0)
@@ -94,7 +94,7 @@ bool SaveAsJsonFile(const std::vector<DumpEntry>& dumpables, const char* filenam
             out << ',';
         }
         out << '[';
-        for (int ii = 0; ii < entry.render_passes.size(); ++ii)
+        for (size_t ii = 0; ii < entry.render_passes.size(); ++ii)
         {
             const DumpRenderPass& render_pass = entry.render_passes[ii];
             if (ii != 0)
@@ -109,7 +109,7 @@ bool SaveAsJsonFile(const std::vector<DumpEntry>& dumpables, const char* filenam
     out << "],\n";
 
     out << "  \"Draw\": [";
-    for (int i = 0; i < dumpables.size(); ++i)
+    for (size_t i = 0; i < dumpables.size(); ++i)
     {
         const DumpEntry& entry = dumpables[i];
         if (i != 0)
@@ -117,7 +117,7 @@ bool SaveAsJsonFile(const std::vector<DumpEntry>& dumpables, const char* filenam
             out << ',';
         }
         out << '[';
-        for (int ii = 0; ii < entry.draws.size(); ++ii)
+        for (size_t ii = 0; ii < entry.draws.size(); ++ii)
         {
             uint64_t draw = entry.draws[ii];
             if (ii != 0)
@@ -131,7 +131,7 @@ bool SaveAsJsonFile(const std::vector<DumpEntry>& dumpables, const char* filenam
     out << "],\n";
 
     out << "  \"QueueSubmit\": [";
-    for (int i = 0; i < dumpables.size(); ++i)
+    for (size_t i = 0; i < dumpables.size(); ++i)
     {
         const DumpEntry& entry = dumpables[i];
         if (i != 0)
